feat: add history and undo/redo

+++ goblinai/server/api/routes/stories.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException
 from goblinai.server.api.schemas.UpdateStorySchema import UpdateStorySchema
 
 from goblinai.server.models.Story import Story
@@ -21,12 +21,19 @@ def postNewStory():
 
 @stories.get("/{id}/")
 def getStoryById(id: str):
-    return Story.getById(id)
+    story = Story.getById(id)
+    if story is None:
+        raise HTTPException(status_code=404, detail="Story not found")
+
+    return story
 
 
 @stories.delete("/{id}/")
 def deleteStoryById(id: str):
     story = Story.getById(id)
+    if story is None:
+        raise HTTPException(status_code=404, detail="Story not found")
+
     story.delete()
     return Story.all()
 
@@ -34,6 +41,9 @@ def deleteStoryById(id: str):
 @stories.patch("/{id}/")
 def updateStory(id: str, body: UpdateStorySchema):
     story = Story.getById(id)
+    if story is None:
+        raise HTTPException(status_code=404, detail="Story not found")
+
     newStory = story.merge(body)
     newStory.save()
     return newStory
